bsi-audit-automator/requirements.txt
# GCP and Vertex AI
google-cloud-storage
google-cloud-aiplatform

# For local development, to load .env files
python-dotenv
jsonschema # For validating AI model outputs